+++ lib/dexter/processor.rb
@@ -11,12 +11,6 @@ module Dexter
       @log_parser =
         if @logfile == :pg_stat_activity
           PgStatActivityParser.new(@indexer, @collector)
-        elsif @logfile == :log_table
-          if options[:input_format] == "csv"
-            CsvLogTableParser.new(@indexer, @collector)
-          else
-            StderrLogTableParser.new(@indexer, @collector)
-          end
         elsif options[:input_format] == "csv"
           CsvLogParser.new(logfile, @collector)
         elsif options[:input_format] == "json"