+++ lib/dexter/stderr_log_table_parser.rb
@@ -1,7 +1,0 @@
-module Dexter
-  class StderrLogTableParser < StderrLogParser
-    def perform
-      process_stderr(@logfile.stderr_activity.map { |r| r["log_entry"] })
-    end
-  end
-end